+++ Common/src/main/java/fuzs/resourcepackoverrides/client/data/ResourceOverridesManager.java
@@ -1,6 +1,7 @@
 package fuzs.resourcepackoverrides.client.data;
 
 import com.google.common.collect.ImmutableList;
+import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Maps;
 import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
@@ -10,6 +11,7 @@ import net.minecraft.network.chat.Component;
 import net.minecraft.server.packs.repository.Pack;
 import net.minecraft.server.packs.repository.PackCompatibility;
 import net.minecraft.util.GsonHelper;
+import org.apache.commons.compress.utils.Lists;
 import org.jetbrains.annotations.Nullable;
 
 import java.io.FileReader;
@@ -21,14 +23,14 @@ import java.util.function.Function;
 public class ResourceOverridesManager {
     private static final String FILE_NAME = ResourcePackOverrides.MOD_ID + ".json";
     private static final String SCHEMA_VERSION = String.valueOf(1);
-    private static final Map<String, PackSelectionOverride> OVERRIDES_BY_ID = Maps.newHashMap();
+    private static Map<String, PackSelectionOverride> overridesById = Maps.newHashMap();
     private static List<String> defaultResourcePacks;
     private static PackSelectionOverride defaultOverride;
     private static int failedReloads;
 
     public static PackSelectionOverride getOverride(String id) {
         if (defaultOverride == null) load();
-        return OVERRIDES_BY_ID.getOrDefault(id, defaultOverride);
+        return overridesById.getOrDefault(id, defaultOverride);
     }
 
     public static List<String> getDefaultResourcePacks(boolean failed) {
@@ -37,14 +39,13 @@ public class ResourceOverridesManager {
         return defaultResourcePacks;
     }
 
-    private static void load() {
+    public static void load() {
         defaultResourcePacks = ImmutableList.of();
         defaultOverride = PackSelectionOverride.EMPTY;
         JsonConfigFileUtil.getAndLoad(FILE_NAME, file -> {}, ResourceOverridesManager::deserializeAllOverrides);
     }
 
     private static void deserializeAllOverrides(FileReader reader) {
-        OVERRIDES_BY_ID.clear();
         JsonElement jsonElement = JsonConfigFileUtil.GSON.fromJson(reader, JsonElement.class);
         JsonObject jsonObject = GsonHelper.convertToJsonObject(jsonElement, "resource pack override");
         if (!GsonHelper.getAsString(jsonObject, "schema_version").equals(SCHEMA_VERSION)) throw new IllegalArgumentException("wrong config schema present");
@@ -61,10 +62,34 @@ public class ResourceOverridesManager {
             defaultOverride = deserializeOverrideEntry(jsonObject.get("default_overrides"));
         }
         if (!jsonObject.has("pack_overrides")) return;
+        Map<String, List<String>> overrideGroups = Maps.newHashMap();
+        if (jsonObject.has("pack_override_groups")) {
+            JsonObject groups = jsonObject.getAsJsonObject("pack_override_groups");
+            for (Map.Entry<String, JsonElement> entry : groups.entrySet()) {
+                JsonArray jsonArray = GsonHelper.convertToJsonArray(entry.getValue(), entry.getKey());
+                List<String> groupIds = overrideGroups.computeIfAbsent(entry.getKey(), id -> Lists.newArrayList());
+                for (JsonElement element : jsonArray) {
+                    groupIds.add(element.getAsString());
+                }
+            }
+        }
+        ImmutableMap.Builder<String, PackSelectionOverride> builder = ImmutableMap.builder();
         JsonObject overrides = jsonObject.getAsJsonObject("pack_overrides");
         for (Map.Entry<String, JsonElement> entry : overrides.entrySet()) {
-            OVERRIDES_BY_ID.put(entry.getKey(), deserializeOverrideEntry(entry.getValue()));
+            JsonElement jsonElement1 = entry.getValue();
+            PackSelectionOverride overrideEntry = deserializeOverrideEntry(jsonElement1);
+            String id = entry.getKey();
+            if (id.startsWith("$")) {
+                List<String> groupIds = overrideGroups.get(id.substring(1));
+                if (groupIds == null) throw new IllegalArgumentException("Unknown group id %s".formatted(id));
+                for (String groupId : groupIds) {
+                    builder.put(groupId, overrideEntry);
+                }
+            } else {
+                builder.put(id, overrideEntry);
+            }
         }
+        overridesById = builder.build();
     }
 
     private static PackSelectionOverride deserializeOverrideEntry(JsonElement jsonElement) {