+++ Forge/src/main/java/fuzs/resourcepackoverrides/ResourcePackOverridesForge.java
@@ -1,8 +1,20 @@
 package fuzs.resourcepackoverrides;
 
+import fuzs.resourcepackoverrides.data.ModLanguageProvider;
+import net.minecraft.data.DataGenerator;
+import net.minecraftforge.common.data.ExistingFileHelper;
+import net.minecraftforge.data.event.GatherDataEvent;
+import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 
 @Mod(ResourcePackOverrides.MOD_ID)
+@Mod.EventBusSubscriber(bus = Mod.EventBusSubscriber.Bus.MOD)
 public class ResourcePackOverridesForge {
 
+    @SubscribeEvent
+    public static void onGatherData(final GatherDataEvent evt) {
+        DataGenerator dataGenerator = evt.getGenerator();
+        final ExistingFileHelper fileHelper = evt.getExistingFileHelper();
+        dataGenerator.addProvider(true, new ModLanguageProvider(dataGenerator, ResourcePackOverrides.MOD_ID));
+    }
 }